refactor: extract MoneyRequestBuilder.ts from IOU/index.ts
Extracts the money-request data-builder cluster (13 functions plus 9 supporting
types) from IOU/index.ts into a new dedicated MoneyRequestBuilder.ts module.

Functions moved:
- buildMinimalTransactionForFormula
- getReportPreviewAction
- getReceiptError
- buildOnyxDataForTestDriveIOU
- getTransactionWithPreservedLocalReceiptSource
- buildOnyxDataForMoneyRequest (~687 lines)
- recalculateOptimisticReportName
- maybeUpdateReportNameForFormulaTitle
- getMoneyRequestInformation (~422 lines)
- calculateDiffAmount
- getUpdatedMoneyRequestReportData
- mergePolicyRecentlyUsedCategories
- mergePolicyRecentlyUsedCurrencies

Function bodies are byte-identical to their pre-move state on main, with the
documented translations of direct module-private state reads
(allReports, allPersonalDetails, allReportActions, allReportNameValuePairs,
deprecatedCurrentUserPersonalDetails, deprecatedUserAccountID) to the
corresponding getters on index.ts (the established pattern from PRs B-D).
Namespace imports (Localize.translateLocal, NumberUtils.rand64) are converted
to named imports (translateLocal, rand64) per lint rules for new files.

A new getCurrentUserPersonalDetails getter is added to index.ts to expose
deprecatedCurrentUserPersonalDetails for the moved functions.

Consumers re-pointed to import from ./MoneyRequestBuilder:
- Split.ts, SplitTransactionUpdate.ts, TrackExpense.ts, PerDiem.ts,
  SendInvoice.ts, Duplicate.ts, BulkEdit.ts, Receipt.ts, PayMoneyRequest.ts,
  DeleteMoneyRequest.ts, UpdateMoneyRequest.ts
- src/pages/inbox/report/ReportActionsView.tsx
- tests/actions/IOU/GetMoneyRequestInformationTest.ts
- tests/actions/IOUTest.ts, IOUTest/SplitTest.ts,
  IOUTest/DeleteMoneyRequestTest.ts, IOUTest/DuplicateTest.ts
- tests/actions/MergeTransactionTest.ts

Final size: index.ts drops from 2,805 to ~1,160 lines.

Part of Issue #72804 (break up src/libs/actions/IOU/).
