+++ src/libs/ReceiptUploadRetryHandler/handleFileRetry.ts
@@ -1,4 +1,5 @@
 import type * as IOU from '@userActions/IOU';
+import type {RequestMoneyInformation} from '@userActions/IOU/MoneyRequestBuilder';
 import {replaceReceipt} from '@userActions/IOU/Receipt';
 import {startSplitBill} from '@userActions/IOU/Split';
 import * as TrackExpense from '@userActions/IOU/TrackExpense';
@@ -6,9 +7,9 @@ import CONST from '@src/CONST';
 import type {ReceiptError} from '@src/types/onyx/Transaction';
 
 export default function handleFileRetry(message: ReceiptError, file: File, dismissError: () => void, setShouldShowErrorModal: (value: boolean) => void) {
-    const retryParams: IOU.ReplaceReceipt | IOU.StartSplitBilActionParams | TrackExpense.CreateTrackExpenseParams | IOU.RequestMoneyInformation =
+    const retryParams: IOU.ReplaceReceipt | IOU.StartSplitBilActionParams | TrackExpense.CreateTrackExpenseParams | RequestMoneyInformation =
         typeof message.retryParams === 'string'
-            ? (JSON.parse(message.retryParams) as IOU.ReplaceReceipt | IOU.StartSplitBilActionParams | TrackExpense.CreateTrackExpenseParams | IOU.RequestMoneyInformation)
+            ? (JSON.parse(message.retryParams) as IOU.ReplaceReceipt | IOU.StartSplitBilActionParams | TrackExpense.CreateTrackExpenseParams | RequestMoneyInformation)
             : message.retryParams;
 
     switch (message.action) {
@@ -38,7 +39,7 @@ export default function handleFileRetry(message: ReceiptError, file: File, dismi
         }
         case CONST.IOU.ACTION_PARAMS.MONEY_REQUEST: {
             dismissError();
-            const requestMoneyParams = {...retryParams} as IOU.RequestMoneyInformation;
+            const requestMoneyParams = {...retryParams} as RequestMoneyInformation;
             requestMoneyParams.transactionParams.receipt = file;
             requestMoneyParams.isRetry = true;
             requestMoneyParams.shouldPlaySound = false;
