+++ src/libs/actions/IOU/BulkEdit.ts
@@ -30,7 +30,8 @@ import ONYXKEYS from '@src/ONYXKEYS';
 import type * as OnyxTypes from '@src/types/onyx';
 import type {SearchResultDataType} from '@src/types/onyx/SearchResults';
 import type {TransactionChanges} from '@src/types/onyx/Transaction';
-import {getAllTransactionViolations, getUpdatedMoneyRequestReportData} from '.';
+import {getAllTransactionViolations} from '.';
+import {getUpdatedMoneyRequestReportData} from './MoneyRequestBuilder';
 
 function removeUnchangedBulkEditFields(
     transactionChanges: TransactionChanges,
