+++ tests/actions/IOU/GetMoneyRequestInformationTest.ts
@@ -1,5 +1,5 @@
 import Onyx from 'react-native-onyx';
-import {getMoneyRequestInformation} from '@libs/actions/IOU';
+import {getMoneyRequestInformation} from '@libs/actions/IOU/MoneyRequestBuilder';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import type {Beta, PolicyTagLists, Report} from '@src/types/onyx';
